+++ Assets/Plugins/Substance 3D Extensions/Asset Extensions/Unity/Timeline/Scripts/Runtime/Substance Track/Clips/Inputs/Core/SetSubstanceInputValueAsset.cs
@@ -30,7 +30,7 @@ namespace SOS.SubstanceExtensions.Timeline
     [System.Serializable]
     public abstract class SetSubstanceInputValueAsset<T> : SetSubstanceInputValueAsset where T : SetSubstanceInputValueBehaviour, new()
     {
-        [SerializeField]
+        [SerializeField, NoFoldout]
         protected T template = new T();
 
         public override ISubstanceInputParameter TargetInputParameter