+++ Assets/Plugins/Substance 3D Extensions/Asset Extensions/Unity/Timeline/Scripts/Runtime/Substance Track/Clips/Render/RenderSubstancesAsset.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+using UnityEngine.Playables;
+
+namespace SOS.SubstanceExtensions.Timeline
+{
+#if UNITY_EDITOR
+    [System.ComponentModel.DisplayName("Render Substance Clip")]
+#endif
+    [System.Serializable]
+    public class RenderSubstancesAsset : PlayableAsset
+    {
+        [SerializeField, NoFoldout]
+        private RenderSubstancesBehaviour template = new RenderSubstancesBehaviour();
+
+        public override Playable CreatePlayable(PlayableGraph graph, GameObject go)
+        {
+            return ScriptPlayable<RenderSubstancesBehaviour>.Create(graph, template);
+        }
+    }
+}