+++ Assets/Plugins/Substance 3D Extensions/Asset Extensions/Unity/Timeline/Scripts/Runtime/Substance Track/SubstanceTrack.cs
@@ -21,6 +21,7 @@ namespace SOS.SubstanceExtensions.Timeline
     [TrackClipType(typeof(SetSubstanceInputInt4Asset))]
     [TrackClipType(typeof(SetSubstanceInputStringAsset))]
     [TrackClipType(typeof(SetSubstanceInputTextureAsset))]
+    [TrackClipType(typeof(RenderSubstancesAsset))]
     public class SubstanceTrack : TrackAsset
     {
         private const string kDisplayNameFormat = "{0} - {1}";