+++ Assets/Plugins/Substance 3D Extensions/Asset Extensions/Unity/Timeline/Scripts/Runtime/Substance Track/SubstanceTrackMixerBehaviour.cs
@@ -347,6 +347,8 @@ namespace SOS.SubstanceExtensions.Timeline
                     if(previousTexture == textures[s]) continue;
                     else previousTextures[s] = textures[s];
 
+                    //playable.GetGraph().Stop();
+
                     _ = drivenNativeGraph.SetInputTextureGPUAsync(s, textures[s]); //Should cache previous value
                 }
             }