+++ Assets/Plugins/Substance 3D Extensions/Scripts/Runtime/Attributes/Utility/NoFoldoutAttribute.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace SOS.SubstanceExtensions
+{
+    /// <summary>
+    /// Attribute that draws classes without the usual foldout.
+    /// </summary>
+    public class NoFoldoutAttribute : PropertyAttribute
+    {
+
+    }
+}